% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{add_above_avg_num}
\alias{add_above_avg_num}
\title{Calculate the rowwise total number for values is higher than column-average value}
\usage{
add_above_avg_num(.data, ..., .name = "above_avg_num", na.rm = TRUE)
}
\arguments{
\item{.data}{data.frame}

\item{...}{columns selected for rowwise `sum` or `mean` using `tidyselect` grammar}

\item{.name}{the name of new column}

\item{na.rm}{whether to delete the missing value, `TRUE` or `FALSE`}
}
\value{

}
\description{
Calculate the rowwise total number for values is higher than column-average value
}
\examples{
df <- tribble(
  ~id, ~x, ~y, ~z, ~g,
  #--|--|--|--|--
  "a", 13.1, 14, 4, 1,
  "b", 11.2, 7, 5, 0,
  "c", 12.5, 10, 1, 0,
  "d", 20, 11, 3, 1
)
df \%>\% add_above_avg_num(x:g)
}
