% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{add_row_means}
\alias{add_row_means}
\title{Calculate the rowwise mean for columns selected in data.frame}
\usage{
add_row_means(.data, ..., .name = "row_mean", na.rm = TRUE)
}
\arguments{
\item{.data}{data.frame}

\item{...}{columns selected for rowwise `sum` or `mean` using `tidyselect` grammar}

\item{.name}{the name of new column}

\item{na.rm}{whether to delete the missing value, `TRUE` or `FALSE`}
}
\value{

}
\description{
Calculate the rowwise mean for columns selected in data.frame
}
\examples{
iris \%>\%
  add_row_means(starts_with("Sepal"), .name = "Sepal.Mean")
}
