% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chengdu}
\alias{chengdu}
\title{Data from student test scores of different schools and classes in chengdu city}
\format{A data frame with 23868 rows and 5 variables:
\describe{
  \item{school}{school name}
  \item{id}{school id}
  \item{class}{class id}
  \item{score_pre}{first test result}
  \item{score_post}{second test result}
}}
\usage{
chengdu
}
\description{
This is data from the two test results from different schools in chengdu city
}
\keyword{datasets}
