% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChineseTones}
\alias{ChineseTones}
\title{speech tones data}
\format{
A data frame with 16 rows and 14 variables:
\describe{
  \item{type}{tone type}
  \item{word}{chinese word}
  \item{p0-p100}{pitch serial}
  \item{time}{speech time}
}
}
\usage{
ChineseTones
}
\description{
speech tones data from experiment using praat, thank you very much for the data set provided by Wu xiaofang.
}
\keyword{datasets}
