% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyuse-package.R
\docType{package}
\name{easyuse-package}
\alias{easyuse}
\alias{easyuse-package}
\title{easyuse: A Package to make social statistics more easy and more tidyverse}
\description{
After a test finished, generally the results from different each school and different class will be put together. From this tidy data frame, the random effects of the test results are assessed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/perlatex/easyuse}
  \item Report bugs at \url{https://github.com/perlatex/easyuse/issues}
}

}
\author{
\strong{Maintainer}: Minjie Wang \email{wangminjiede@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
