% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{five_step_standard}
\alias{five_step_standard}
\title{standardized a numeric vector using five step way}
\usage{
five_step_standard(x)
}
\arguments{
\item{.x}{a numeric vector}
}
\value{
a standardized vector
}
\description{
standardized formula is `(x - min(x)) * 5 / (max(x) - min(x))`
}
\examples{
mtcars \%>\%
  mutate(fstep = five_step_standard(mpg))
}
