% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ran_vals.R
\name{get_ran_vals}
\alias{get_ran_vals}
\title{Assess Random-Effect of Student Test Scores}
\usage{
get_ran_vals(.data, .var_school, .var_class, .var_score_pre,
  .var_score_after, effects = "class")
}
\arguments{
\item{.data}{data.frame including school name, class name, score-pre and score-after of each student}

\item{.var_school}{school name.}

\item{.var_class}{class name.}

\item{.var_score_pre}{score-pre.}

\item{.var_score_after}{score-after, which is used as the response variable in mixed models}

\item{effects}{specification for effects, which is either class or school}
}
\description{
From this tidy data frame, the random effects of the test results are assessed using lme4::lmer funciton.
}
\examples{
get_ran_vals(df, school, class, score1, score2, "class")

}
