% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_first_match.R
\name{row_first_match}
\alias{row_first_match}
\title{Find the value of the first match in rowwise}
\usage{
row_first_match(.f, ...)
}
\arguments{
\item{.f}{A function, formula, or vector, like in `purrr::detect()`

  If a function, it is used as is.
  
  If a formula, e.g. ~ .x + 2, it is converted to a function. There
  are three ways to refer to the arguments:


  * For a single argument function, use `.`
     
  * For a two argument function, use `.x` and `.y`
     
  * For more arguments, use `..1`, `..2`, `..3` etc
  
  This syntax allows you to create very compact anonymous functions.}

\item{...}{column names selected in data.frame, e.g. `x, y, z`.}
}
\value{

}
\description{
Find the value of the first match in rowwise. The code from https://community.rstudio.com/t/interesting-tidy-eval-use-cases/21121/39
}
\examples{
df <- tibble(
  a = c("b", "d", "l", "m"), 
  x = c(1, 1, 1, 2), 
  y = c(5, 1, 2, 3), 
  z = c(1, 1, 0, 1)
)


df \%>\% dplyr::mutate(new_col = first_match(~ . > 1, x, y, z))
}
